requirements.txt
Py3DNS
dnspython